--- idaes_ui/fv/api/__init__.py
@@ -1,1 +1,0 @@
-from .diag_models import DiagnosticsData, DiagnosticsError
