+++ idaes_ui/fv/models/diag.py
@@ -1,0 +1,38 @@
+"""
+Data model for combined IDAES model diagnostic data
+"""
+__author__ = "Dan Gunter"
+__created__ = "2023-10-08"
+
+# external packages
+from pydantic import BaseModel, computed_field
+# IDAES/Pyomo
+from pyomo.core.base.block import _BlockData
+from idaes.core.util import model_diagnostics as imd
+# package
+from .stats import ModelStats
+from .issues import ModelIssues
+
+
+# Data model
+
+class DiagnosticsData(BaseModel):
+    """The standard set of diagnostics data for a flowsheet"""
+
+    config: dict = dict(imd.CONFIG)
+
+    def __init__(self, block: _BlockData):
+        super().__init__()
+        self._ms = ModelStats(block)
+        self._mi = ModelIssues(block)
+
+    @computed_field
+    @property
+    def statistics(self) -> ModelStats:
+        return self._ms
+
+    @computed_field
+    @property
+    def issues(self) -> ModelIssues:
+        self._mi.update()
+        return self._mi
