+++ idaes_ui/fv/models/issues.py
@@ -7,15 +7,11 @@ __author__ = "Dan Gunter"
 __created__ = "2023-09-05"
 
 # stdlib
-import argparse
 from enum import Enum
-import json
-import sys
 from typing import Dict, Union, List, Tuple
 
 # third-party
-from pydantic import BaseModel, computed_field
-from pandas import DataFrame
+from pydantic import BaseModel
 
 # Import IDAES functionality (use shortened names, i=IDAES m=model)
 from idaes.core.util import model_statistics as ims
@@ -26,149 +22,19 @@ from pyomo.core.base.block import _BlockData
 from pyomo.environ import value
 from pyomo.util.check_units import identify_inconsistent_units
 
+from .base import DiagnosticsException
 
-# Wrap model_statistics module
-# ============================
-
-
-class StatisticsUpdateError(Exception):
-    def __init__(self, details=None):
-        msg = f"Statistics update failed"
-        if details:
-            msg += f": {details}"
-        super().__init__(msg)
+# Exceptions
 
 
-# Pydantic model for statistics
-# -----------------------------
-
-
-class StatsCount(BaseModel):
-    # Value
-    value: int = 0
-
-
-class StatsCountDeactType(StatsCount):
-    deact: StatsCount = StatsCount()
-
-
-class StatsCountVarFixed(StatsCount):
-    unused: StatsCount = StatsCount()
-
-
-class StatsCountVarIneq(StatsCount):
-    fixed: StatsCount = StatsCount()
-
-
-class StatsCountConstraints(StatsCount):
-    eq: StatsCountDeactType = StatsCountDeactType()
-    ineq: StatsCountDeactType = StatsCountDeactType()
-
-
-class StatsCountVar(StatsCount):
-    unused: StatsCount = StatsCount()
-    fixed: StatsCountVarFixed = StatsCountVarFixed()
-    ineq: StatsCountVarIneq = StatsCountVarIneq()
-
-
-class ModelStats(BaseModel):
-    """Model Statistics API
-
-    For each named item (or sub item), access the value as ".value"::
-
-        m = ModelStats()
-        m.dof.value = 3
-        m.var.fixed.ineq.value = 5
+class DiagnosticsUpdateException(DiagnosticsException):
+    """Error updating diagnostics
     """
-
-    dof: StatsCount = StatsCount()
-    var: StatsCountVar = StatsCountVar()
-    ineq: StatsCount = StatsCount()
-    constr: StatsCountConstraints = StatsCountConstraints()
-    obj: StatsCountDeactType = StatsCountDeactType()
-    block: StatsCountDeactType = StatsCountDeactType()
-    expr: StatsCount = StatsCount()
-
-    def __init__(self, block: _BlockData):
-        super().__init__()
-        self._block = block
-        self.update()
-
-    def update(self):
-        """Update values in report."""
-        b = self._block
-        try:
-            # variables
-            self.dof.value = ims.degrees_of_freedom(b)
-            self.var.value = ims.number_variables(b)
-            self.var.fixed.value = ims.number_fixed_variables(b)
-            self.var.unused.value = ims.number_unused_variables(b)
-            self.var.fixed.unused.value = ims.number_fixed_unused_variables(b)
-            self.var.ineq.value = ims.number_variables_only_in_inequalities(b)
-            self.var.ineq.fixed.value = ims.number_fixed_variables_only_in_inequalities(
-                b
-            )
-            # constraints / objectives
-            self.constr.value = ims.number_total_constraints(b)
-            self.constr.eq.value = ims.number_total_equalities(b)
-            self.constr.eq.deact.value = ims.number_deactivated_equalities(b)
-            self.constr.ineq.value = ims.number_total_inequalities(b)
-            self.constr.ineq.deact.value = ims.number_deactivated_inequalities(b)
-            self.obj.value = ims.number_total_objectives(b)
-            self.obj.deact.value = ims.number_deactivated_objectives(b)
-            # blocks
-            self.block.value = ims.number_total_blocks(b)
-            self.block.deact.value = ims.number_deactivated_blocks(b)
-            # expressions
-            self.expr.value = ims.number_expressions(b)
-        except Exception as err:
-            raise StatisticsUpdateError(details=str(err))
-
-    def as_table(self) -> DataFrame:
-        """Transform model information into a table.
-
-        Returns:
-            DataFrame: Pandas DataFrame with the following columns:
-
-                1. type = Main type, e.g. 'var' for variables
-                2. subtypes = Comma-separated list of subtypes, e.g. 'fixed,unused'.
-                   If empty, this is the total for the type.
-                3. attr = For convenience, the type + subtypes as a dot-separated
-                   attribute. Simply add ".value" to fetch value
-                4. value = Count for this combination of type and subtype(s)
-        """
-        a = []
-        for type_, info in self.model_dump().items():
-            for subtypes, val in self._as_table_subtypes(info, ()):
-                attr_name = ".".join([type_] + list(subtypes))
-                a.append((type_, ",".join(subtypes), attr_name, val))
-        return DataFrame(data=a, columns=["type", "subtypes", "attr", "value"])
-
-    @classmethod
-    def _as_table_subtypes(
-        cls, info: Dict[str, Union[Dict, float]], subtypes: Tuple[str] | Tuple
-    ) -> List[Tuple[Tuple[str], float]]:
-        result = []
-        for key in info:
-            if key == "value":
-                result.append((subtypes, info[key]))
-            else:
-                nested_types = tuple(list(subtypes) + [key])  # for type-checker
-                nested_result = cls._as_table_subtypes(info[key], nested_types)
-                result.extend(nested_result)
-        return result
+    def __init__(self, name="unknown", details=None):
+        super().__init__(name=name, details=details)
 
 
-# Wrap model_diagnostics module
-# =============================
-
-
-class DiagnosticsError(Exception):
-    def __init__(self, name, details=None):
-        msg = f"Diagnostic '{name}' failed"
-        if details:
-            msg += f": {details}"
-        super().__init__(msg)
+# Diagnostics
 
 
 class Severity(str, Enum):
@@ -404,65 +270,3 @@ class ModelIssues(BaseModel):
         self.issues.append(issue)
 
 
-class DiagnosticsData(BaseModel):
-    """The standard set of diagnostics data for a flowsheet"""
-
-    config: dict = dict(imd.CONFIG)
-
-    def __init__(self, block: _BlockData):
-        super().__init__()
-        self._ms = ModelStats(block)
-        self._mi = ModelIssues(block)
-
-    @computed_field
-    @property
-    def statistics(self) -> ModelStats:
-        return self._ms
-
-    @computed_field
-    @property
-    def issues(self) -> ModelIssues:
-        self._mi.update()
-        return self._mi
-
-
-def print_sample_output(output_file, indent=False, solve=True, schema=False):
-    from idaes_ui.fv.tests.flowsheets import idaes_demo_flowsheet
-
-    flowsheet = idaes_demo_flowsheet()
-    if solve:
-        flowsheet.solve()
-    data = DiagnosticsData(flowsheet)
-
-    output_file.write(data.model_dump_json(indent=indent))
-
-    if schema:
-        output_file.write("\n----\n")
-        json.dump(
-            data.model_json_schema(mode="serialization"), output_file, indent=indent
-        )
-
-
-if __name__ == "__main__":
-    p = argparse.ArgumentParser()
-    p.add_argument("-f", "--file", help="Output file (default=stdout)")
-    p.add_argument("-i", "--indent", help="indent JSON", action="store_true")
-    p.add_argument("-n", "--no-solve", help="do not solve model", action="store_true")
-    p.add_argument("-s", "--schema", help="also print JSON schema", action="store_true")
-    args = p.parse_args()
-    # choose output stream
-    if args.file:
-        try:
-            ofile = open(args.file, "w")
-        except IOError as e:
-            print(f"Failed to open output file '{args.file}': {e}")
-            sys.exit(-1)
-    else:
-        ofile = sys.stdout
-
-    # create output
-    print_sample_output(
-        ofile, indent=args.indent, solve=not args.no_solve, schema=args.schema
-    )
-    if ofile is not sys.stdout:
-        print(f"\nWrote output to file: {args.file}")
